moved supervisor's citizen-related behavior to its own module
