+++ lib/modules/supervisor/citizen/notify-on-crash-or-shutdown.js
