+++ lib/supervisor.js
@@ -1,6 +1,6 @@
 var merge = require('merge-objects'),
     Noticeboard = require('cjs-noticeboard'),
-    middleware = require('../utils/middleware');
+    Middleware = require('../utils/middleware');
 
 module.exports = supe;
 
@@ -9,12 +9,12 @@ function supe( config ){
   var citizens = {},
       config = merge({ retries: 3, duration: 3 }, config || {} ),      
       noticeboard = new Noticeboard({ logging: false, logOps: false }),
-      my_middleware = middleware(),
+      middleware = Middleware(),
       public_interface = {};
 
   // create supervisor interface
     public_interface.noticeboard = noticeboard;
-    public_interface.middleware = my_middleware;
+    public_interface.middleware = middleware;
 
     public_interface.register = register_citizen;
     public_interface.start = start_citizen;
@@ -22,9 +22,7 @@ function supe( config ){
     public_interface.use = load_module;
 
   // load core modules
-    public_interface.use( require( './modules/supervisor/pipe-citizen-output' ) );
-    public_interface.use( require( './modules/supervisor/notify-on-crash-or-shutdown' ) );
-    public_interface.use( require( './modules/supervisor/restart-crashed-citizens' ) );
+    public_interface.use( require( './modules/supervisor/citizen' ) );
     public_interface.use( require( './modules/supervisor/signal' ) );
     public_interface.use( require( './modules/supervisor/mail' ) );
